requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
fastmcp>=0.2.0
starlette>=0.27.0 